core/requirements.txt
django
django-paypal
logging
uuid
os
sys
pathlib
